python-app/requirements.txt
psycopg2-binary
SQLAlchemy==1.4.39
Flask==2.1.1
gunicorn==20.1.0
python-dotenv==0.20.0
flask